Move simulator to Dialect/LLHD/Simulator
