+++ lib/CMakeLists.txt
@@ -2,5 +2,4 @@ add_subdirectory(Conversion)
 add_subdirectory(Dialect)
 add_subdirectory(FIRParser)
 add_subdirectory(EmitVerilog)
-add_subdirectory(Simulator)
 add_subdirectory(Target)
