+++ lib/Dialect/LLHD/CMakeLists.txt
@@ -1,2 +1,3 @@
 add_subdirectory(IR)
+add_subdirectory(Simulator)
 add_subdirectory(Transforms)
