+++ lib/Dialect/LLHD/Simulator/Engine.cpp
@@ -1,6 +1,6 @@
-#include "circt/Simulator/Engine.h"
+#include "circt/Dialect/LLHD/Simulator/Engine.h"
 #include "circt/Conversion/LLHDToLLVM/LLHDToLLVM.h"
-#include "circt/Simulator/State.h"
+#include "circt/Dialect/LLHD/Simulator/State.h"
 
 #include "mlir/ExecutionEngine/ExecutionEngine.h"
 #include "mlir/IR/Module.h"
